refactoring funcs/examples as hydra-reference
